Clarifications

+++ inc/Standards.hpp
@@ -4,7 +4,7 @@
 // Some stuff you can control for the entire library
 namespace RTB {
 
-using RESOLUTION = float;
+using RESOLUTION = double;
 
 // Slightly unnecessary - use cmath or math.h instead?
 constexpr RESOLUTION PI = 3.141592653589793238462643383279502884;
@@ -13,6 +13,18 @@ constexpr RESOLUTION PI_2 = PI / 2;
 constexpr RESOLUTION PI_4 = PI / 4;
 constexpr RESOLUTION PI_180 = PI / 180;
 constexpr RESOLUTION R180_PI = 180 / PI;
+
+// Axes follow the SOFA convention:
+// https://www.sofaconventions.org/mediawiki/index.php/SOFA_specifications
+// Cartesian is thus basically the physics convention (right hand rule)
+// FRONT = +X, BACK = -X
+// LEFT = +Y, RIGHT = -Y
+// UP = +Z, DOWN = -Z
+// Spherical just has the Elevation reversed
+// +Azimuth = rotate from +X to +Y
+// -Azimuth = rotate from +X to -Y
+// +Elevation = rotate from XY-Plane to +Z
+// -Elevation = rotate from XY-Plane to -Z
 
 }  // namespace RTB
 #endif  // STANDARDS_HPP
